refactor(v010): carve voicegw rotate-secret out of _legacy
Second per-command carve-out in section 2.

Adds voicegateway/cli/rotate_secret.py owning the 'rotate-secret'
command (re-encrypts managed_providers rows under a new primary
Fernet key while still decrypting under the previous one held in
VOICEGW_SECRET_FALLBACK). The command body is ~85 LOC, self-contained
modulo _load_gateway which still lives in _legacy.py and is imported
explicitly.

Removes 'import os' from _legacy.py: rotate-secret was the only
remaining consumer (os.environ lookups for VOICEGW_SECRET and
VOICEGW_SECRET_FALLBACK). ruff would flag F401 otherwise.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package.
